+++ cx-portal/src/components/pages/ServiceMarketplaceDetail/components/MarketplaceHeader/index.tsx
@@ -28,7 +28,6 @@ import { OVERLAYS, ROLES } from 'types/Constants'
 import UserService from 'services/UserService'
 import './MarketplaceHeader.scss'
 import { setSuccessType } from 'features/serviceMarketplace/slice'
-import { getServiceLeadImage } from 'features/apps/mapper'
 
 export default function MarketplaceHeader({
   item,
@@ -76,7 +75,7 @@ export default function MarketplaceHeader({
   return (
     <div className="marketplace-header">
       <div className="lead-image">
-        <img src={getServiceLeadImage(item)} alt={item.title} />
+        <img src={'/ServiceMarketplace.png'} alt={item.title} />
       </div>
       <div className="content">
         <Typography variant="body2" className="provider">