+++ CHANGELOG.md
@@ -4,7 +4,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 
 
 ### Unreleased
-n/a
+* Feature - Service marketplace UI Updates
 
 ## 0.9.0